# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Glory Sports API - Cricket sports ecosystem for India
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /players:
    get:
      operationId: listPlayers
      tags: [players]
      summary: List and discover players
      parameters:
        - name: city
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            type: string
        - name: minRating
          in: query
          schema:
            type: number
        - name: available
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of players
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerListResponse"
  /players/{playerId}:
    get:
      operationId: getPlayer
      tags: [players]
      summary: Get player profile
      parameters:
        - name: playerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Player profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerProfile"
  /players/{playerId}/stats:
    get:
      operationId: getPlayerStats
      tags: [players]
      summary: Get player report card and stats
      parameters:
        - name: playerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Player stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerStats"
  /players/me/profile:
    put:
      operationId: updateMyProfile
      tags: [players]
      summary: Update own player profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerProfileInput"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerProfile"
  /players/me/availability:
    put:
      operationId: updateAvailability
      tags: [players]
      summary: Update player availability status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvailabilityInput"
      responses:
        "200":
          description: Updated availability
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerProfile"

  # NEED PLAYERS POSTS
  /need-players:
    get:
      operationId: listNeedPlayers
      tags: [players]
      summary: List "need players" posts
      parameters:
        - name: city
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Need players posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NeedPlayersPost"
    post:
      operationId: createNeedPlayersPost
      tags: [players]
      summary: Post a need players request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NeedPlayersInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NeedPlayersPost"
  /need-players/{postId}/join:
    post:
      operationId: joinNeedPlayersPost
      tags: [players]
      summary: Join a need players post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Joined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NeedPlayersPost"

  # TEAMS
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List user notifications
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{notificationId}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notification marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"

  # ADMIN
...
components:
  schemas:
...
    PlayerProfile:
      type: object
      required: [id, userId, name]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        city:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        playerRole:
          type: ["string", "null"]
          enum: [batsman, bowler, all_rounder, wicket_keeper, null]
        battingStyle:
          type: ["string", "null"]
        bowlingStyle:
          type: ["string", "null"]
        availabilityStatus:
          type: ["string", "null"]
          enum: [available_today, available_weekend, looking_for_team, unavailable, null]
        overallRating:
          type: ["number", "null"]
        battingRating:
          type: ["number", "null"]
        bowlingRating:
          type: ["number", "null"]
        isVerified:
          type: boolean
        trustScore:
          type: ["number", "null"]
        fairPlayScore:
          type: ["number", "null"]
        verifiedMatchPct:
          type: ["number", "null"]
        teamsCount:
          type: integer
        matchesCount:
          type: integer
    PlayerProfileInput:
      type: object
      properties:
        city:
          type: string
        playerRole:
          type: string
          enum: [batsman, bowler, all_rounder, wicket_keeper]
        battingStyle:
          type: string
        bowlingStyle:
          type: string
        avatarUrl:
          type: string
    PlayerStats:
      type: object
      required: [playerId]
      properties:
        playerId:
          type: integer
        totalMatches:
          type: integer
        totalRuns:
          type: integer
        totalWickets:
          type: integer
        strikeRate:
          type: ["number", "null"]
        economy:
          type: ["number", "null"]
        highestScore:
          type: ["integer", "null"]
        winPercentage:
          type: ["number", "null"]
        mvpCount:
          type: integer
        verifiedMatchCount:
          type: integer
        recentMatches:
          type: array
          items:
            $ref: "#/components/schemas/MatchSummary"
    PlayerListResponse:
      type: object
      required: [players, total]
      properties:
        players:
          type: array
          items:
            $ref: "#/components/schemas/PlayerProfile"
        total:
          type: integer
    AvailabilityInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [available_today, available_weekend, looking_for_team, unavailable]
    NeedPlayersPost:
      type: object
      required: [id, userId, title, city, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        city:
          type: string
        sport:
          type: string
        neededCount:
          type: integer
        joinedCount:
          type: integer
        matchDate:
          type: ["string", "null"]
        createdAt:
          type: string
    NeedPlayersInput:
      type: object
      required: [title, city, neededCount]
      properties:
        title:
          type: string
        description:
          type: string
        city:
          type: string
        sport:
          type: string
        neededCount:
          type: integer
        matchDate:
          type: string

    # TEAMS
...
    MatchSummary:
      type: object
      required: [matchId, matchDate, result]
      properties:
        matchId:
          type: integer
        matchDate:
          type: string
        result:
          type: string
        teamA:
          type: string
        teamB:
          type: string
        myTeamScore:
          type: ["integer", "null"]
        opponentScore:
          type: ["integer", "null"]
        isMvp:
          type: boolean

    # STATS / LEADERBOARD
...
    Notification:
      type: object
      required: [id, userId, type, title, isRead, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
        title:
          type: string
        body:
          type: ["string", "null"]
        isRead:
          type: boolean
        referenceId:
          type: ["integer", "null"]
        referenceType:
          type: ["string", "null"]
        createdAt:
          type: string

    # ADMIN
...