# lib/api-zod/src/generated/types/playerProfileInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Glory Sports API - Cricket sports ecosystem for India
 * OpenAPI spec version: 0.1.0
 */
import type { PlayerProfileInputPlayerRole } from "./playerProfileInputPlayerRole";

export interface PlayerProfileInput {
  city?: string;
  playerRole?: PlayerProfileInputPlayerRole;
  battingStyle?: string;
  bowlingStyle?: string;
  avatarUrl?: string;
}